feat: replace account deactivation email job with new job for multiple days and update cron job initialization
